refactor: auth 스토어 모듈 경로 재구성 및 임포트 경로 업데이트\n- useAuthStore를 '@/src/shared/store/auth-store' 로 이동시켜 모듈 경로 통합- page.tsx와 auth-provider.tsx의 임포트 경로 수정하여 새 위치 참조- src/entities/.gitkeep 삭제로 불필요한 빈 파일 제거
