+++ src/shared/store/auth-store/index.ts
