+++ src/shared/store/auth-store/useAuthStore.ts
